add some level of message abstraction for minion
