+++ hydra-filters/src/main/java/com/addthis/hydra/data/filter/bundle/BundleFilter.java
@@ -109,10 +109,19 @@ public abstract class BundleFilter implements Codec.Codable {
                 } finally {
                     init.set(true);
                     initing.set(false);
+                    synchronized (this) {
+                        this.notifyAll();
+                    }
                 }
             } else {
                 while (initing.get()) {
-                    Thread.yield();
+                    try {
+                        synchronized (this) {
+                            this.wait(100);
+                        }
+                    } catch (InterruptedException e) {
+                        throw new RuntimeException(e);
+                    }
                 }
             }
         }
